+++ crates/miden-objects/src/transaction/proven_tx.rs
@@ -141,21 +141,21 @@ impl ProvenTransaction {
     /// - The size of the serialized account update exceeds [`ACCOUNT_UPDATE_MAX_SIZE`].
     /// - The transaction is empty, which is the case if the account state is unchanged or the
     ///   number of input notes is zero.
-    /// - The transaction was executed against a _new_ on-chain account and its account ID does not
-    ///   match the ID in the account update.
-    /// - The transaction was executed against a _new_ on-chain account and its commitment does not
-    ///   match the final state commitment of the account update.
+    /// - The transaction was executed against a _new_ account with public state and its account ID
+    ///   does not match the ID in the account update.
+    /// - The transaction was executed against a _new_ account with public state and its commitment
+    ///   does not match the final state commitment of the account update.
     /// - The transaction was executed against a private account and the account update is _not_ of
     ///   type [`AccountUpdateDetails::Private`].
-    /// - The transaction was executed against an on-chain account and the update is of type
-    ///   [`AccountUpdateDetails::Private`].
-    /// - The transaction was executed against an _existing_ on-chain account and the update is of
-    ///   type [`AccountUpdateDetails::New`].
-    /// - The transaction creates a _new_ on-chain account and the update is of type
+    /// - The transaction was executed against an account with public state and the update is of
+    ///   type [`AccountUpdateDetails::Private`].
+    /// - The transaction was executed against an _existing_ account with public state and the
+    ///   update is of type [`AccountUpdateDetails::New`].
+    /// - The transaction creates a _new_ account with public state and the update is of type
     ///   [`AccountUpdateDetails::Delta`].
     fn validate(self) -> Result<Self, ProvenTransactionError> {
-        // If the account is on-chain, then the account update details must be present.
-        if self.account_id().is_onchain() {
+        // If the account's state is public, then the account update details must be present.
+        if self.account_id().has_public_state() {
             self.account_update.validate()?;
 
             // check that either the account state was changed or at least one note was consumed,
@@ -170,14 +170,14 @@ impl ProvenTransaction {
             let is_new_account = self.account_update.initial_state_commitment() == Word::empty();
             match self.account_update.details() {
                 AccountUpdateDetails::Private => {
-                    return Err(ProvenTransactionError::OnChainAccountMissingDetails(
+                    return Err(ProvenTransactionError::PublicStateAccountMissingDetails(
                         self.account_id(),
                     ));
                 },
                 AccountUpdateDetails::New(account) => {
                     if !is_new_account {
                         return Err(
-                            ProvenTransactionError::ExistingOnChainAccountRequiresDeltaDetails(
+                            ProvenTransactionError::ExistingPublicStateAccountRequiresDeltaDetails(
                                 self.account_id(),
                             ),
                         );
@@ -197,9 +197,11 @@ impl ProvenTransaction {
                 },
                 AccountUpdateDetails::Delta(_) => {
                     if is_new_account {
-                        return Err(ProvenTransactionError::NewOnChainAccountRequiresFullDetails(
-                            self.account_id(),
-                        ));
+                        return Err(
+                            ProvenTransactionError::NewPublicStateAccountRequiresFullDetails(
+                                self.account_id(),
+                            ),
+                        );
                     }
                 },
             }
@@ -380,17 +382,17 @@ impl ProvenTransactionBuilder {
     /// - The size of the serialized account update exceeds [`ACCOUNT_UPDATE_MAX_SIZE`].
     /// - The transaction is empty, which is the case if the account state is unchanged or the
     ///   number of input notes is zero.
-    /// - The transaction was executed against a _new_ on-chain account and its account ID does not
-    ///   match the ID in the account update.
-    /// - The transaction was executed against a _new_ on-chain account and its commitment does not
-    ///   match the final state commitment of the account update.
+    /// - The transaction was executed against a _new_ account with public state and its account ID
+    ///   does not match the ID in the account update.
+    /// - The transaction was executed against a _new_ account with public state and its commitment
+    ///   does not match the final state commitment of the account update.
     /// - The transaction was executed against a private account and the account update is _not_ of
     ///   type [`AccountUpdateDetails::Private`].
-    /// - The transaction was executed against an on-chain account and the update is of type
-    ///   [`AccountUpdateDetails::Private`].
-    /// - The transaction was executed against an _existing_ on-chain account and the update is of
-    ///   type [`AccountUpdateDetails::New`].
-    /// - The transaction creates a _new_ on-chain account and the update is of type
+    /// - The transaction was executed against an account with public state and the update is of
+    ///   type [`AccountUpdateDetails::Private`].
+    /// - The transaction was executed against an _existing_ account with public state and the
+    ///   update is of type [`AccountUpdateDetails::New`].
+    /// - The transaction creates a _new_ account with public state and the update is of type
     ///   [`AccountUpdateDetails::Delta`].
     pub fn build(self) -> Result<ProvenTransaction, ProvenTransactionError> {
         let input_notes =