+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - [BREAKING] Incremented MSRV to 1.89.
 - [BREAKING] Remove versioning of the transaction kernel, leaving only one latest version ([#1793](https://github.com/0xMiden/miden-base/pull/1793)).
+- [BREAKING] Rename the `is_onchain` method to `has_public_state` for `AccountId`, `AccountIdPrefix`, `Account`, `AccountInterface` and `AccountStorageMode` ([#1846](https://github.com/0xMiden/miden-base/pull/1846)).
 
 ## 0.11.1 (2025-08-28)